scripts/pipeline/requirements.txt
httpx==0.27.2
feedparser==6.0.11
python-slugify==8.0.4
tenacity==9.0.0

